+++ api/tests/models/artworks.test.js
@@ -3,3 +3,10 @@
 // write controller methods
 // write test files for controller methods
 // create routes for artwork
+
+require("../mongodb_helper");
+const Artwork = require("../../models/artwork")
+
+describe("Artwork model constructs correctly", () => {
+  
+})
