+++ frontend/src/assets/styles/MainBar.css
