+++ frontend/src/assets/styles/NavBar.css
