+++ frontend/src/assets/styles/SideBar.css
